v64 Update
Changes to the moons to make them v64 compatible. I know it's been a while, I've been busy and v55 to v64 was too short and killed my motivation to update, but with the game's updates being halted I finally got around to updating the project and moons. More mods will be coming in the future. :)